% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterRound.R
\name{betterRound}
\alias{betterRound}
\title{Round a number, preserving extra 0s}
\usage{
betterRound(x, digits = 2L)
}
\arguments{
\item{x}{Vector of numbers to round.}

\item{digits}{Number of digits past the decimal point to keep.}
}
\value{
A character vector of rounded numbers.
}
\description{
Round a number, preserving extra 0s.
}
\details{
Uses \code{\link[base]{sprintf}} to round a number, keeping extra 0s.
}
\examples{
betterRound(51.01, 3)
betterRound(0.199, 2)
}
