% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{between}
\alias{between}
\alias{\%btwn\%}
\alias{between}
\title{Between}
\usage{
x \%btwn\% range

between(x, range, inclusive = TRUE)
}
\arguments{
\item{x}{Numeric value or vector to check}

\item{range}{Numeric vector of two elements, defining the range to check}

\item{inclusive}{Logical indicating whether or not to include the lower and upper
values of the range [Default = TRUE]}
}
\value{
a logical vector of the same length as \code{x}.  For every element in \code{x} returns
\code{TRUE} if the value lies between the range and \code{FALSE} if not.
}
\description{
Determine if a numeric value is within a range
}
\details{
\code{x \%btwn\% range} is equivalent to \code{beween(x, range, inclusive = TRUE)}
}
\examples{
5 \%btwn\% c(1, 10)
c(5, 20) \%btwn\% c(5, 10)
between(5, c(5, 10))
between(5, c(5, 10), inclusive = FALSE)

}
