\name{check_jags_code}
\alias{check_jags_code}
\title{Check JAGS model code}
\usage{
check_jags_code(x)
}
\arguments{
  \item{x}{string of JAGS model code}
}
\value{
Invisible flag of whether JAGS model code passes certain
checks. In addition, a unique warning is issued for each
failed check.
}
\description{
Checks JAGS model code
}
\seealso{
\code{\link{juggler}}
}

