% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDir.R
\name{createDir}
\alias{createDir}
\title{Create directories}
\usage{
createDir(dir.names)
}
\arguments{
\item{dir.names}{character vector of the directory paths to be created}
}
\value{
returns invisibly a logical vector indicating if the
  operation succeeded for each of the files attempted. Using a missing
  value for a path name will always be regarded as a failure.
  \code{createDir} will will fail if the directory already exists.
}
\description{
Wrapper around the base function \code{dir.create} that allows creation
 of multiple directories simultaneously in a recursive fashion without
 warning messages
}
\examples{
\dontrun{
createDir("testdir")
}
}
