% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfCharacterize.R
\name{dfCharacterize}
\alias{dfCharacterize}
\title{'Characterize' data.frame}
\usage{
dfCharacterize(df, only = c(), ignore = c())
}
\arguments{
\item{df}{A data.frame.}

\item{only}{A vector of column names. Only convert these columns to characters.}

\item{ignore}{A vector of column names. Do not convert these columns.}
}
\value{
A data.frame with all factor columns (or only a subset of them
  if the parameters were set) converted to characters.

a data.frame with factor columns converted to characters.
}
\description{
Convert factor columns in a data.frame to characters.
}
\details{
Given a data.frame, convert factor columns to characters.  By default,
all factor columns are converted, but the user can specify which columns
to convert or to not convert.
}
\note{
If any of the names in \code{only} or \code{ignore} are not valid
  columns, an error is raised.

Only one of \code{only} or \code{ignore} can be used in a single call.
}
\examples{
df <- data.frame(a = 1:3, b = letters[1:3], c = LETTERS[1:3],
                 stringsAsFactors = TRUE)
str(df)
str(dfCharacterize(df))
str(dfCharacterize(df, only = "b"))
str(dfCharacterize(df, ignore = "b"))
str(dfCharacterize(df, ignore = c("b", "c")))
str(dfCharacterize(df, only = c("a")))

# The following examples result in errors
\dontrun{str(dfCharacterize(df, only = c("z")))}
\dontrun{str(dfCharacterize(df, only = "b", ignore = "c"))}
}
