% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveFile.R
\name{file.move}
\alias{file.move}
\title{Move a file}
\usage{
file.move(from, to)
}
\arguments{
\item{from}{path to file to move}

\item{to}{path to destination}
}
\value{
Logical vector indicating if operation successful
}
\description{
Move a file from one location to another and create destination directories
as necessary
}
