% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolowerfirst.R
\name{firstcharacter}
\alias{firstcharacter}
\alias{tolowerfirst}
\alias{toupperfirst}
\title{Convert first character to lower or upper case}
\usage{
tolowerfirst(x)

toupperfirst(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
The original character vector with the first character of every
  element converted to lower or upper case
}
\description{
Given a character vector, convert the first character to lower or upper case.
}
\details{
\code{tolowerfirst} converts the first character in a string to lower case, and
\code{toupperfirst} converts the first character in a string to lower case.
}
\examples{
tolowerfirst("CamelCase")
tolowerfirst("ALLCAPS")
tolowerfirst(c("First", "_Second"))
toupperfirst("camelcase")
toupperfirst("allcaps")
toupperfirst(c("first", "_second"))

}
