% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag_class.R
\docType{data}
\name{frag_class}
\alias{frag_class}
\title{Model Code Fragment class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
frag_class
}
\description{
Model Code Fragment class
}
\section{Fields}{

\describe{
\item{\code{whole}}{entire model code fragment}

\item{\code{variable_name}}{name of variable being assigned to. Includes square brackets and contents}

\item{\code{operator}}{string containing the operator for the model code fragment}

\item{\code{expression}}{string containing the entire expression to be assigned to the variable_name}

\item{\code{comment}}{string containing the comment (or command string for models and predictions)}
}}
\keyword{datasets}

