% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-model-code-version.r
\name{generate_model_code_version}
\alias{generate_model_code_version}
\title{Generate a single version of a Model_Code}
\usage{
generate_model_code_version(x, fragments)
}
\arguments{
\item{x}{is Super Model Code to replicate}

\item{fragments}{is Fragments defining different versions of the Model}
}
\description{
Generate a single version of a Model_Code
}

