% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jg_ModelCode.R
\name{is.jg_ModelCode}
\alias{is.jg_ModelCode}
\title{Is jg_ModelCode}
\usage{
is.jg_ModelCode(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class \code{\link{jg_ModelCode}}.
}

