% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jg_ModelCodeFragment.R
\name{is.jg_ModelCodeFragment}
\alias{is.jg_ModelCodeFragment}
\title{Is jg_ModelCodeFragment}
\usage{
is.jg_ModelCodeFragment(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class \code{\link{jg_ModelCodeFragment}}.
}

