% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserved.R
\name{jags_block_names}
\alias{jags_block_names}
\title{JAGS Block Names}
\usage{
jags_block_names()
}
\value{
A character vector of the permitted blocks names in
their required order.
}
\description{
JAGS Block Names
}

