\name{jg_block_names}
\alias{jg_block_names}
\title{Get and set JAGS model code block names}
\usage{
jg_block_names(x)
}
\arguments{
  \item{x}{string of JAGS model code}
}
\value{
Character vector of block names.
}
\description{
Gets character vector of the names of the blocks in a JAGS
model code string
}
\seealso{
\code{\link{juggler}}, \code{\link{jg_nblocks}} and
\code{\link{jg_set_block}}
}

