\name{jg_data}
\alias{jg_data}
\title{Get data variables in JAGS model code}
\usage{
jg_data(x)
}
\arguments{
  \item{x}{string of JAGS model code}
}
\value{
Character vector of data variables.
}
\description{
Gets data variables in JAGS model code
}
\seealso{
\code{\link{juggler}}, \code{\link{jg_dists}},
\code{\link{jg_funcs}}, \code{\link{jg_parms}} and
\code{\link{jg_indices}}
}

