\name{jg_dists}
\alias{jg_dists}
\title{Get distributions in JAGS model code}
\usage{
jg_dists(x)
}
\arguments{
  \item{x}{string of JAGS model code}
}
\value{
Character vector of distributions.
}
\description{
Gets distributions in JAGS model code
}
\seealso{
\code{\link{juggler}}, \code{\link{jg_funcs}},
\code{\link{jg_parms}} \code{\link{jg_indices}} and
\code{\link{jg_data}}
}

