% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-node-lines.R
\name{jg_extract_modelcode_fragment}
\alias{jg_extract_modelcode_fragment}
\title{Parse Node Lines}
\usage{
jg_extract_modelcode_fragment(x)
}
\arguments{
\item{x}{to parse.}
}
\description{
Parse Node Lines
}

