\name{jg_funcs}
\alias{jg_funcs}
\title{Get functions in JAGS model code}
\usage{
jg_funcs(x)
}
\arguments{
  \item{x}{string of JAGS model code}
}
\value{
Character vector of functions.
}
\description{
Gets functions in JAGS model code
}
\seealso{
\code{\link{juggler}}, \code{\link{jg_dists}},
\code{\link{jg_parms}} \code{\link{jg_indices}} and
\code{\link{jg_data}}
}

