% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jg_generate_modelcode.r
\name{jg_generate_modelcode}
\alias{jg_generate_modelcode}
\title{Generate a single version of a Model_Code}
\usage{
jg_generate_modelcode(x, fragments)
}
\arguments{
\item{x}{is Super Model Code to replicate}

\item{fragments}{is Fragments defining different versions of the Model}
}
\description{
Generate a single version of a Model_Code
}

