% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juggle.R
\name{jg_juggle}
\alias{jg_juggle}
\title{Juggle JAGS model code}
\usage{
jg_juggle(x, fragments = character(0))
}
\arguments{
\item{x}{string of JAGS model code}

\item{fragments}{A character vector of model fragments}
}
\value{
A character vector of alternative JAGS models.
}
\description{
Juggles JAGS model code, i.e., 
takes code in the JAGS dialect of the BUGS language
and produces alternative models from model fragments.
}
\details{
jg_juggle currently requires that each node be defined on
a separate line of code. It takes advantage of the fact
that each node is unique.
}

