% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jg_ModelCode.R
\docType{data}
\name{jg_ModelCode}
\alias{jg_ModelCode}
\title{Model Code class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
jg_ModelCode
}
\description{
Model Code class
}
\section{Fields}{

\describe{
\item{\code{model}}{Contains a string representation of the vmodel version of the Model Code}

\item{\code{prediction}}{Contains a string representation of the vprediction version of the Model Code}
}}
\keyword{datasets}

