\name{jg_nblocks}
\alias{jg_nblocks}
\title{Get number of JAGS model code blocks}
\usage{
jg_nblocks(x)
}
\arguments{
  \item{x}{string of JAGS model code}
}
\value{
Count of the number of blocks.
}
\description{
Gets the number of blocks in a JAGS model code string
}
\seealso{
\code{\link{juggler}}, \code{\link{jg_blocks}} and
\code{\link{jg_set_block}}
}

