% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jg_replicate_modelcode.R
\name{jg_replicate_modelcode}
\alias{jg_replicate_modelcode}
\title{Generate every version of a Model_Code}
\usage{
jg_replicate_modelcode(x, fragments)
}
\arguments{
\item{x}{is Super Model Code to replicate}

\item{fragments}{is a vector of Fragments defining different versions of the Model}
}
\description{
Generate every version of a Model_Code
}

