% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jg_specialise_modelcode.R
\name{jg_specialise_modelcode}
\alias{jg_specialise_modelcode}
\title{Generate the prediction and model versions of the Model_Code}
\usage{
jg_specialise_modelcode(x)
}
\arguments{
\item{x}{is Super Model Code to replicate}
}
\description{
Generate the prediction and model versions of the Model_Code
}

