% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgmodel.R
\name{jgmodel}
\alias{jgmodel}
\title{JG Model}
\usage{
jgmodel(model_code)
}
\arguments{
\item{model_code}{A string or character vector of the model code.}
}
\value{
A jgmodel object
}
\description{
JG Model
}

