% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_code_class.R
\docType{data}
\name{model_code_class}
\alias{model_code_class}
\title{Model Code class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model_code_class
}
\description{
Model Code class
}
\section{Fields}{

\describe{
\item{\code{model}}{Contains a string representation of the vmodel version of the Model Code}

\item{\code{prediction}}{Contains a string representation of the vprediction version of the Model Code}
}}
\keyword{datasets}

