% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{number}
\alias{number}
\alias{addCommas}
\alias{addSpaces}
\alias{percent}
\alias{compPercent}
\title{Number formatters}
\usage{
number(x, accuracy = 1, multiplier = 1, prefix = "", suffix = "",
  big.mark = " ", decimal.mark = ".")

addCommas(x, accuracy = 1)

addSpaces(x, accuracy = 1)

percent(x, accuracy = 1)

compPercent(x, accuracy = 1)
}
\arguments{
\item{x}{a numeric vector to format}

\item{accuracy}{number to round to, \code{NULL} for automatic guess.
[10 wll round to nearest 10, 1 will round to nearest integer, 0.1 to nearest 0.1, etc.]}

\item{multiplier}{number to multiply by (e.g. for computing percentages or thousands)}

\item{prefix, suffix}{Symbols to display before and after value}

\item{big.mark}{character used between every 3 digits to separate thousands}

\item{decimal.mark}{the character to be used to indicate the numeric decimal point}
}
\value{
a formatted character vector
}
\description{
\code{number} is a generic formatter for numeric values.
\code{addCommas} is a shortcut for comma as separator for thousands, point for decimal.
\code{addSpaces} is a shortcut for space as separator for thousands, point for decimal.
\code{percent} a shortcut for percentages (value are multiplied by 100 and a % symbol is added).
\code{compPercent} returns the complement of 1, i.e. \code{percent(1 - x)}.
}
\examples{
v <- c(12.3, 4, 12345.789, 0.0002)
number(v)
addCommas(v, accuracy = 0.01)
p <- runif(10)
percent(p)
compPercent(p)
}
