% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSize.R
\name{objectSize}
\alias{objectSize}
\title{Object sizes}
\usage{
objectSize(decreasing = TRUE, head = FALSE, n = 10)
}
\arguments{
\item{decreasing}{If \code{TRUE}, sort the objects from largest to smallest.}

\item{head}{If \code{TRUE}, only show the top \code{n} objects.}

\item{n}{Integer specifying the number of values to show if \code{head = TRUE}.}
}
\value{
A data.frame with the name, class, and size of each object. If the object
is a data.frame or matrix, it will also specify its dimensions.
}
\description{
Calculate the sizes of objects in the workspace.
}
\examples{
print(output <- objectSize())

}
