% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDir.R
\name{openDir}
\alias{openDir}
\title{Open a directory in file browser}
\usage{
openDir(dir = getwd())
}
\arguments{
\item{dir}{The absolute or relative path to the directory to be opened}
}
\value{
Invisible NULL.  The functions puts up a window and returns immediately.
  The window can be closed via its controls or menus.
}
\description{
\code{openDir} opens the specified directory in the
  default file browser.  When no directory is specified
  the current working directory is opened.
}
\note{
This has only been tested on Windows
}
\examples{
\dontrun{
openDir()
openDir("..")
}
}
