% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFile.R
\name{openFile}
\alias{openFile}
\title{Open a file}
\usage{
openFile(file)
}
\arguments{
\item{file}{Path to file to be opened}
}
\value{
Invisble NULL.  The function opens the file using its default program and
 returns immediately.  The window cam be closed via its controls or menus.
}
\description{
Open a file using the default program.
}
\examples{
\dontrun{openFile("myplot.pdf")}
}
