% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\alias{signifStars}
\title{p-values formatter and significance stars}
\usage{
pval(x, accuracy = 0.001, stars = FALSE, three = 0.001, two = 0.01,
  one = 0.05)

signifStars(x, three = 0.001, two = 0.01, one = 0.05)
}
\arguments{
\item{x}{A numeric vector of p-values}

\item{accuracy}{Number to round to [Default = 0.001]}

\item{stars}{If \code{TRUE} then add significance stars [Default = FALSE]}

\item{three}{Number indicating level below which to display three stars '***' [Default = 0.001]}

\item{two}{Number indicating level below which to display two stars '**' [Default = 0.01]}

\item{one}{Number indicating level below which to display one star '*' [Default = 0.05]}
}
\value{
A character vector
}
\description{
Formatter for p-values, adding a symbol "<" for very small p-values and,
optionally, significance stars
}
\examples{
p <- c(.50, 0.12, .09, .045, .011, .009, .00002, NA)
pval(p)
pval(p, accuracy = .01)
pval(p, stars = TRUE)
signifStars(p)
signifStars(p, one = .15)
}
