% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTrail.R
\name{removeTrail}
\alias{removeTrail}
\title{Remove trailing white spaces}
\usage{
removeTrail(str, before = TRUE, after = TRUE, char = " ")
}
\arguments{
\item{str}{character string to be trimmed}

\item{before}{a logical which indicates whether to remove the
character from the beginning of the string}

\item{after}{a logical which indicates whether to remove the
character from the end of the string}

\item{char}{the character to be removed}
}
\value{
Returns a string from which the character specified has been removed.
  If the argument \code{char} is not specified, then this will be trailing
  white spaces.
}
\description{
Remove trailing white spaces (or other character) from a string
}
\examples{
str <- "  this is a test   "
x <- removeTrail(str, before = TRUE)
x <- removeTrail(str, after = TRUE)
x <- removeTrail(str, before = TRUE, after = TRUE)
x <- removeTrail(str, before = TRUE, char = "  th")
}
