% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate-model-codes.r
\name{replicate_model_codes}
\alias{replicate_model_codes}
\title{Replicate Model_Codes}
\usage{
replicate_model_codes(x, fragments)
}
\arguments{
\item{x}{is Super Model Code to replicate}

\item{fragments}{is Fragments defining different versions of the Model}
}
\description{
Replicate Model_Codes
}

