% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialise_model_and_prediction.R
\name{specialise_model_and_prediction}
\alias{specialise_model_and_prediction}
\title{Generate the prediction and model versions of the Model_Code}
\usage{
specialise_model_and_prediction(x)
}
\arguments{
\item{x}{is Super Model Code to replicate}
}
\description{
Generate the prediction and model versions of the Model_Code
}

