% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_stack.R
\docType{data}
\name{state_stack_class}
\alias{state_stack_class}
\title{Stack Classs for storing a stack of states class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
state_stack_class
}
\description{
Stack Classs for storing a stack of states class
}
\keyword{datasets}

