% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringify.R
\name{stringify}
\alias{stringify}
\title{Convert a vector into a single character string}
\usage{
stringify(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A character string
}
\description{
Convert a vector into a single character string with the items
separated by commas and an "and".
}
\examples{
stringify(letters[1:2])
stringify(letters[1:4])
}
