% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedf.R
\name{write.df}
\alias{write.df}
\title{Write a data.frame to file with sensible defaults}
\usage{
write.df(x, file = "", quote = FALSE, sep = "\\t",
  row.names = FALSE, col.names = TRUE)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame.
    If not, it is attempted to coerce \code{x} to a data frame.}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing.  \code{""} indicates output to the console.}

\item{quote}{a logical value (\code{TRUE} or \code{FALSE}) or a
    numeric vector.  If \code{TRUE}, any character or factor columns
    will be surrounded by double quotes.  If a numeric vector, its
    elements are taken as the indices of columns to quote.  In both
    cases, row and column names are quoted if they are written.  If
    \code{FALSE}, nothing is quoted.}

\item{sep}{the field separator string.  Values within each row of
    \code{x} are separated by this string.}

\item{row.names}{either a logical value indicating whether the row
    names of \code{x} are to be written along with \code{x}, or a
    character vector of row names to be written.}

\item{col.names}{either a logical value indicating whether the column
    names of \code{x} are to be written along with \code{x}, or a
    character vector of column names to be written.  See the section on
    \sQuote{CSV files} for the meaning of \code{col.names = NA}.}
}
\description{
This is simply a wrapper around write.table with the
settings I use most often: quote = FALSE, sep = "\t", row.names = FALSE,
and col.names = TRUE
}
\seealso{
\code{\link{utils}{write.table}}
}
