% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{jg_vnodes}
\alias{jg_vnodes}
\title{Get variable nodes in JAGS model code}
\usage{
jg_vnodes(x, type = "both", indices = FALSE)
}
\arguments{
\item{x}{string of JAGS model code}

\item{type}{string of node type. Must be 'stochastic', 'deterministic'
or 'both' (the default). A variable node is any word possibly}

\item{indices}{flag of whether to retain indices}
}
\value{
Character vector of unique sorted variable node names
}
\description{
Gets names of variable (as opposed to constant) named nodes in JAGS model code
}
\seealso{
\code{\link{juggler}}, \code{\link{jg_dists}},
and \code{\link{jg_funcs}}
}

