% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_model_codes.R
\name{replicate_model_codes}
\alias{replicate_model_codes}
\title{Generate every version of a Model_Code}
\usage{
replicate_model_codes(x, fragments)
}
\arguments{
\item{x}{is Super Model Code to replicate}

\item{fragments}{is a vector of Fragments defining different versions of the Model}
}
\description{
Generate every version of a Model_Code
}

